later

+++ code/modules/atmospherics/environmental/LINDA_fire.dm
@@ -24,24 +24,47 @@
 					active_hotspot.volume = exposed_volume
 		return
 
-	if((exposed_temperature > PLASMA_MINIMUM_BURN_TEMPERATURE) && has_fuel)
+	if((exposed_temperature > FIRE_MINIMUM_TEMPERATURE_TO_EXIST) && has_fuel)
+
+
 		active_hotspot = new /obj/effect/hotspot(src, exposed_volume*25, exposed_temperature)
 
-//This is the icon for fire on turfs, also helps for nurturing small fires until they are full tile
+		/*
+		active_hotspot.just_spawned = (current_cycle < SSair.times_fired)
+			//remove just_spawned protection if no longer processing this cell
+		SSair.add_to_active(src)
+		*/
+
+/**
+ * Hotspot objects interfaces with the temperature of turf gasmixtures while also providing visual effects.
+ * One important thing to note about hotspots are that they can roughly be divided into two categories based on the bypassing variable.
+ */
 /obj/effect/hotspot
 	anchored = TRUE
 	mouse_opacity = MOUSE_OPACITY_TRANSPARENT
 	icon = 'icons/effects/fire.dmi'
 	icon_state = "1"
 	layer = GASFIRE_LAYER
+	plane = ABOVE_GAME_PLANE
 	blend_mode = BLEND_ADD
 	light_system = MOVABLE_LIGHT
 	light_range = LIGHT_RANGE_FIRE
 	light_power = 1
 	light_color = LIGHT_COLOR_FIRE
 
+	/**
+	 * Volume is the representation of how big and healthy a fire is.
+	 * Hotspot volume will be divided by turf volume to get the ratio for temperature setting on non bypassing mode.
+	 * Also some visual stuffs for fainter fires.
+	 */
 	var/volume = 125
+	/// Temperature handles the initial ignition and the colouring.
 	var/temperature = FIRE_MINIMUM_TEMPERATURE_TO_EXIST
+	/*
+	/// Whether the hotspot is new or not. Used for bypass logic.
+	var/just_spawned = TRUE
+	*/
+	/// Whether the hotspot becomes passive and follows the gasmix temp instead of changing it.
 	var/bypassing = FALSE
 	var/visual_update_tick = 0
 	var/first_cycle = TRUE
@@ -62,6 +85,19 @@
 	)
 	AddElement(/datum/element/connect_loc, loc_connections)
 
+/**
+ * Perform interactions between the hotspot and the gasmixture.
+ *
+ * For the first tick, hotspots will take a sample of the air in the turf,
+ * set the temperature equal to a certain amount, and then reacts it.
+ * In some implementations the ratio comes out to around 1, so all of the air in the turf.
+ *
+ * Afterwards if the reaction is big enough it mostly just tags along the fire,
+ * copying the temperature and handling the colouring.
+ * If the reaction is too small it will perform like the first tick.
+ *
+ * Also calls fire_act() which handles burning.
+ */
 /obj/effect/hotspot/proc/perform_exposure()
 	var/turf/open/location = loc
 	if(!istype(location) || !(location.air))
@@ -91,6 +127,7 @@
 			AT.fire_act(temperature, volume)
 	return
 
+/// Mathematics to be used for color calculation.
 /obj/effect/hotspot/proc/gauss_lerp(x, x1, x2)
 	var/b = (x1 + x2) * 0.5
 	var/c = (x2 - x1) / 6
@@ -150,7 +187,20 @@
 	alpha = heat_a
 
 #define INSUFFICIENT(path) (location.air.get_moles(path) < 0.5)
+
+/**
+ * Regular process proc for hotspots governed by the controller.
+ * Handles the calling of perform_exposure() which handles the bulk of temperature processing.
+ * Burning or fire_act() are also called by perform_exposure().
+ * Also handles the dying and qdeletion of the hotspot and hotspot creations on adjacent cardinal turfs.
+ * And some visual stuffs too! Colors and fainter icons for specific conditions.
+ */
 /obj/effect/hotspot/process()
+	/*
+	if(just_spawned)
+		just_spawned = FALSE
+		return
+	*/
 	var/turf/open/location = loc
 	if(!istype(location))
 		qdel(src)
@@ -220,7 +270,6 @@
 
 /obj/effect/hotspot/proc/on_entered(datum/source, atom/movable/arrived, atom/old_loc, list/atom/old_locs)
 	SIGNAL_HANDLER
-
 	if(isliving(arrived))
 		var/mob/living/immolated = arrived
 		immolated.fire_act(temperature, volume)
